✨ wps打卡版适配青龙
青龙目前仅支持渠道1打卡
